+++ src/Controller/EmployeeController.php
@@ -37,12 +37,16 @@ class EmployeeController extends AbstractController
     }
 
     /**
-     * @Route("admin_api/createemployee/{name}")
+     * @Route("admin_api/createemployee")
      */
-    public function createEmployee($name,  LoggerInterface $logger, Request $request,EmployeeApi $employeeApi): Response
+    public function createEmployee(LoggerInterface $logger, Request $request,EmployeeApi $employeeApi): Response
     {
         $logger->info("Starting Method: " . __METHOD__);
-        $response = $employeeApi->createEmployee($name);
+        if (!$request->isMethod('post')) {
+            return new JsonResponse("Internal server error" , 500, array());
+        }
+
+        $response = $employeeApi->createEmployee($request->get('name'));
         $callback = $request->get('callback');
         $response = new JsonResponse($response , 200, array());
         $response->setCallback($callback);
@@ -56,6 +60,9 @@ class EmployeeController extends AbstractController
     public function deleteEmployee($employeeId, LoggerInterface $logger, Request $request,EmployeeApi $employeeApi): Response
     {
         $logger->info("Starting Method: " . __METHOD__);
+        if (!$request->isMethod('remove')) {
+            return new JsonResponse("Internal server error" , 500, array());
+        }
         $response = $employeeApi->deleteEmployee($employeeId);
         $callback = $request->get('callback');
         $response = new JsonResponse($response , 200, array());
@@ -69,6 +76,9 @@ class EmployeeController extends AbstractController
     public function updateEmployees($employeeId, $newValue, LoggerInterface $logger,Request $request, EntityManagerInterface $entityManager, EmployeeApi $employeeApi): Response
     {
         $logger->info("Starting Method: " . __METHOD__);
+        if (!$request->isMethod('put')) {
+            return new JsonResponse("Internal server error" , 500, array());
+        }
         $response = $employeeApi->updateEmployeeName($employeeId, $newValue);
         $callback = $request->get('callback');
         $response = new JsonResponse($response , 200, array());
